requirements.txt
numpy==2.1.3
pandas==2.2.3
matplotlib==3.9.2
plotly==5.24.1
dash==3.0.0
dash-bootstrap-components==2.0.0
seaborn==0.13.2
python-dotenv==1.0.1
diskcache==5.6.3
psutil==7.0.0
multiprocess==0.70.17
aiohttp==3.11.14
asyncio==3.4.3